refactor: split refresh execution into dedicated file
