+++ config
@@ -22,7 +22,7 @@ fi
 
 # Module name and source files
 ngx_addon_name=ngx_http_cache_purge_module
-CACHE_PURGE_SRCS="$ngx_addon_dir/ngx_cache_purge_module.c"
+CACHE_PURGE_SRCS="$ngx_addon_dir/ngx_cache_purge_module.c $ngx_addon_dir/ngx_http_cache_refresh.c"
 
 # Build configuration for different nginx versions
 if [ -n "$ngx_module_link" ]; then
